refactor(transport): extract types + constants from androidTvRemote.ts
